+++ src/main/java/domain/JanggiBoardFactory.java
@@ -1,5 +1,15 @@
 package domain;
 
+import domain.piece.Empty;
+import domain.piece.Piece;
+import domain.piece.Side;
+import domain.piece.궁;
+import domain.piece.마;
+import domain.piece.사;
+import domain.piece.상;
+import domain.piece.졸병;
+import domain.piece.차;
+import domain.piece.포;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
