+++ src/main/java/domain/piece/PieceStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum PieceStatus {
     ACTIVE,
