+++ src/main/java/domain/piece/Side.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum Side {
     초,
