+++ src/test/java/domain/piece/PieceTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
